refactor: split builtin_codecs.py into package
Split the 1,286-line builtin_codecs.py into a builtin_codecs/ package
with separate modules for each codec:

- blob.py: BlobCodec (Python object serialization)
- hash.py: HashCodec (hash-addressed storage)
- schema.py: SchemaCodec (base class for schema-addressed codecs)
- object.py: ObjectCodec (files/folders)
- attach.py: AttachCodec (file attachments)
- filepath.py: FilepathCodec (references to existing files)
- npy.py: NpyCodec + NpyRef (numpy arrays)

Benefits:
- Each codec is now self-contained and easier to understand
- Easier to maintain and test individual codecs
- Clearer organization for contributors
- No change to public API (same imports work)
